requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.3.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
PyPDF2==3.0.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
soupsieve==2.5
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
